+++ app/routes/components/TimeSelector/demos/Demo01.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 
 import { type TimeRange, TimeSelector } from "src/components/TimeSelector";
 
-export function Demo() {
+export default function Demo() {
   const [timeRange, setTimeRange] = useState<TimeRange | undefined>(undefined);
 
   return <TimeSelector value={timeRange} onChange={setTimeRange} />;