+++ src/components/TimeSelector/index.ts
@@ -1,4 +1,4 @@
-// TimeSelector from copy-ui @ 2025-10-04.
+// TimeSelector from copy-ui @ 2025-11-06.
 
-export type { TimeRange } from "./TimeSelector";
-export { TimeSelector } from "./TimeSelector";
+export type { TimeRange } from "./time-selector";
+export { TimeSelector } from "./time-selector";