+++ src/components/TimeSelector/CHANGELOG.md
@@ -5,3 +5,5 @@
 - 2025-10-04: Replace preset list with ScrollArea component. Add proper form
   labels with `htmlFor` attributes. Extract preset data to array. Update button
   styles and cancel button variant. Code formatting improvements.
+- 2025-11-06: Refactor code and use `Field` component for input fields. Adjust
+  input widths and spacing, etc.